+++ src/wagtailmarkdown/mdx/linkers/document.py
@@ -1,20 +1,6 @@
-# vim:sw=4 ts=4 et:
-# Copyright (c) 2015 Torchbox Ltd.
-# felicity@torchbox.com 2015-09-14
-#
-# Permission is granted to anyone to use this software for any purpose,
-# including commercial applications, and to alter it and redistribute it
-# freely. This software is provided 'as-is', without any express or implied
-# warranty.
-#
-
 from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
 
-
-try:  # wagtail < 2.0
-    from wagtail.wagtaildocs.models import Document
-except ImportError:  # wagtail >= 2.0
-    from wagtail.documents.models import Document
+from wagtail.documents.models import Document
 
 from markdown.util import etree