+++ src/wagtailmarkdown/mdx/linkers/image.py
@@ -1,21 +1,8 @@
-# vim:sw=4 ts=4 et:
-# Copyright (c) 2015 Torchbox Ltd.
-# felicity@torchbox.com 2015-09-14
-#
-# Permission is granted to anyone to use this software for any purpose,
-# including commercial applications, and to alter it and redistribute it
-# freely. This software is provided 'as-is', without any express or implied
-# warranty.
-#
 from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
 
+from wagtail.images import get_image_model
+
 from markdown.util import etree
-
-
-try:  # wagtail < 2.0
-    from wagtail.wagtailimages import get_image_model
-except ImportError:  # wagtail >= 2.0
-    from wagtail.images import get_image_model
 
 
 # TODO: Default spec and class should be configurable, because they're