+++ src/wagtailmarkdown/mdx/linkers/page.py
@@ -1,21 +1,9 @@
-# vim:sw=4 ts=4 et:
-# Copyright (c) 2015 Torchbox Ltd.
-# felicity@torchbox.com 2015-09-14
-#
-# Permission is granted to anyone to use this software for any purpose,
-# including commercial applications, and to alter it and redistribute it
-# freely. This software is provided 'as-is', without any express or implied
-# warranty.
-#
 from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
+
+from wagtail.core.models import Page
 
 from markdown.util import etree
 
-
-try:
-    from wagtail.core.models import Page
-except ImportError:
-    from wagtail.wagtailcore.models import Page
 
 # TODO: In Waiflike, this only allowed linking to SitePage (the main
 # content type).  Should this be configurable?