+++ src/wagtailmarkdown/warnings.py
@@ -1,13 +1,2 @@
-# vim:sw=4 ts=4 et:
-# Copyright (c) 2015 Torchbox Ltd.
-# tomasz.knapik@torchbox.com 2017-12-07
-#
-# Permission is granted to anyone to use this software for any purpose,
-# including commercial applications, and to alter it and redistribute it
-# freely. This software is provided 'as-is', without any express or implied
-# warranty.
-#
-
-
 class WagtailMarkdownDeprecationWarning(DeprecationWarning):
     pass